requirements.txt
streamlit>=1.20.0
tensorflow>=2.10.0
numpy
pillow
